Move diff artifacts to one dir
