--- diff-models/core/.editorconfig
@@ -1,15 +1,0 @@
-root = true
-
-[*]
-charset = utf-8
-end_of_line = lf
-trim_trailing_whitespace = true
-insert_final_newline = true
-
-[*.md]
-insert_final_newline = false
-trim_trailing_whitespace = false
-
-[*.{js,json,ts,yml,yaml}]
-indent_size = 2
-indent_style = space
