--- diff-models/core/.eslintrc.json
@@ -1,27 +1,0 @@
-{
-  "env": {
-    "browser": false,
-    "es6": true,
-    "node": true
-  },
-  "parser": "@typescript-eslint/parser",
-  "parserOptions": {
-    "project": "tsconfig.json",
-    "sourceType": "module",
-    "ecmaVersion": 2020
-  },
-  "plugins": [
-    "@typescript-eslint",
-    "jest"
-  ],
-  "extends": [
-    "eslint:recommended",
-    "plugin:@typescript-eslint/eslint-recommended",
-    "plugin:@typescript-eslint/recommended",
-    "plugin:jest/recommended",
-    "prettier"
-  ],
-  "rules": {
-    "@typescript-eslint/no-var-requires": 0
-  }
-}
