--- diff-models/core/.prettierrc
@@ -1,17 +1,0 @@
-{
-  "singleQuote": true,
-  "trailingComma": "all",
-  "overrides": [
-    {
-      "files": [
-        "*.ts",
-        "*.mts"
-      ],
-      "options": {
-        "parser": "typescript"
-      }
-    }
-  ],
-  "tabWidth": 2,
-  "printWidth": 74
-}
