--- diff-models/core/__tests__/tsconfig.json
@@ -1,30 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2022",
-    "module": "node16",
-    "lib": [
-      "ES2022"
-    ],
-    "moduleResolution": "Node16",
-    "rootDir": "..",
-    "outDir": "build",
-    "allowSyntheticDefaultImports": true,
-    "importHelpers": true,
-    "alwaysStrict": true,
-    "sourceMap": true,
-    "forceConsistentCasingInFileNames": true,
-    "noFallthroughCasesInSwitch": true,
-    "noImplicitReturns": true,
-    "noUnusedLocals": false,
-    "noUnusedParameters": false,
-    "noImplicitAny": true,
-    "noImplicitThis": true,
-    "strictNullChecks": true,
-    "allowJs": true,
-    "esModuleInterop": true
-  },
-  "include": [
-    "src/**/*",
-    "__tests__/**/*"
-  ]
-}
