--- diff-models/core/jest.config.js
@@ -1,24 +1,0 @@
-export default {
-  testEnvironment: 'node',
-  preset: 'ts-jest/presets/js-with-ts-esm',
-  globals: {
-    'ts-jest': {
-      useESM: true,
-      tsconfig: '<rootDir>/__tests__/tsconfig.json',
-    },
-  },
-  transformIgnorePatterns: [
-    "node_modules/(?!(time-span|convert-hrtime))",
-  ],
-  moduleNameMapper: {
-    '^(\\.{1,2}/.*)\\.(m)?js$': '$1',
-  },
-  testRegex: '(/__tests__/.*|(\\.|/)(test|spec))\\.(m)?ts$',
-  coverageDirectory: 'coverage',
-  collectCoverageFrom: [
-    'src/**/*.ts',
-    'src/**/*.mts',
-    '!src/**/*.d.ts',
-    '!src/**/*.d.mts',
-  ],
-};
