--- diff-models/core/package.json
@@ -1,50 +1,0 @@
-{
-  "name": "node-typescript-boilerplate",
-  "version": "0.0.0",
-  "description": "Minimalistic boilerplate to quick-start Node.js development in TypeScript.",
-  "type": "module",
-  "engines": {
-    "node": ">= 16.13 <17"
-  },
-  "devDependencies": {
-    "@types/jest": "^28.1.4",
-    "@types/node": "~16",
-    "@typescript-eslint/eslint-plugin": "~5.26",
-    "@typescript-eslint/parser": "~5.26",
-    "eslint": "~8.16",
-    "eslint-config-prettier": "~8.5",
-    "eslint-plugin-jest": "~26.2",
-    "jest": "^28.1.1",
-    "prettier": "~2.6",
-    "rimraf": "~3.0",
-    "source-map-support": "^0.5.21",
-    "ts-jest": "^28.0.5",
-    "tsutils": "~3.21",
-    "typescript": "~4.7"
-  },
-  "scripts": {
-    "start": "node build/src/main.js",
-    "clean": "rimraf coverage build tmp",
-    "prebuild": "npm run lint",
-    "build": "tsc -p tsconfig.json",
-    "build:watch": "tsc -w -p tsconfig.json",
-    "build:release": "npm run clean && tsc -p tsconfig.release.json",
-    "lint": "eslint . --ext .ts --ext .mts",
-    "test": "jest --coverage",
-    "prettier": "prettier --config .prettierrc --write .",
-    "test:watch": "jest --watch"
-  },
-  "author": "Jakub Synowiec <jsynowiec@users.noreply.github.com>",
-  "license": "Apache-2.0",
-  "dependencies": {
-    "@types/clone-deep": "^4.0.1",
-    "@types/underscore": "^1.11.4",
-    "clone-deep": "^4.0.1",
-    "time-span": "^5.1.0",
-    "tslib": "~2.4",
-    "underscore": "^1.13.4"
-  },
-  "volta": {
-    "node": "16.13.0"
-  }
-}
