--- diff-models/core/tsconfig.release.json
@@ -1,8 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "sourceMap": false,
-    "removeComments": true
-  },
-  "include": ["src/**/*"]
-}
