--- diff-tests/core/.gitignore
@@ -1,3 +1,0 @@
-# *.json
-*.log
-help.py
