+++ diff-tests/core/driver/.gitignore
@@ -1,0 +1,1 @@
+*.log
