+++ diff-tests/core/model/.eslintignore
@@ -1,0 +1,1 @@
+/**/*.js
