--- diff-tests/core/traces.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:4b0891f64c8f757921983cbf38e40a19cb95e9b9c35f102e1a8acc604b67ce03
-size 5280557
